Add stack equality and unit tests for stacks
This commit adds equality checking to the stack implementation
(`stack`). This is implemented in the function `stack_equal`
(in `src/stack.c`).

This commit also adds unit tests for the stack implementation.

This commit also adds these unit tests, and the unit tests for the
doubly-linked list implementation (`dlist`) to the build system
(in `Makefile`).